[의정] 상세페이지 기능 구현
[의정] 상세페이지 기능 구현

+++ migrations/20220505084957-create-restaurant.js
@@ -32,7 +32,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT(3,2)
       },
       parking: {
         type: Sequelize.BOOLEAN
